test(mcp): Share structured content assertions
Move structuredContent test helpers into a shared test utility so structured result tests use one contract for extracting payloads and checking generated fallback text.
